requirements.txt
rdkit-pypi
biopython
pandas
scikit-learn
torch
transformers
tqdm
